+++ mb-url-http.el
@@ -117,6 +117,40 @@ This function deletes the first block (from proxy)."
           (while (re-search-forward "\r\n" nil t)
             (replace-match "\n")))))))
 
+(defun mb-url-http--fix-header (header fix-function &optional last all list)
+  "Fix all HEADER lines from response message.
+
+FIX-FUNCTION should be a function of one argument.  It is called with header
+values as its argument and returns the new value of HEADER.  If it returns a
+string, the value of HEADER will be replaced with the string.  If it returns
+nil, all HEADER lines will be removed.  If it returns t, all HEADER lines
+keeps unchanged.
+
+LAST, ALL, LIST will be passed to `mail-fetch-field'.
+
+This function is only implemented in Emacs 27+."
+  (when (< emacs-major-version 27)
+    (error "Not implemented in Emacs %s" emacs-version))
+  (save-excursion
+    (save-restriction
+      (goto-char (point-min))
+      (and (not (buffer-narrowed-p))
+           (mail-narrow-to-head))
+      (goto-char (point-min))
+      (let* ((hvals (mail-fetch-field header last all list nil))
+             (fixed (funcall fix-function hvals)))
+        (cond ((eq fixed t))
+              ((null fixed)
+               (goto-char (point-min))
+               (mail-fetch-field header last all list t)
+               (flush-lines "^[ \t]*$" (point-min) (point-max)))
+              (t
+               (goto-char (point-min))
+               (mail-fetch-field header last all list t)
+               (flush-lines "^[ \t]*$" (point-min) (point-max))
+               (goto-char (point-max))
+               (insert (concat header ": " fixed "\n"))))))))
+
 (defun mb-url-http--reset-end-of-headers ()
   "Reset url-http-end-of-headers."
   (setq url-http-end-of-headers
@@ -300,7 +334,9 @@ If SENTINEL is nil, `mb-url-http-sentinel' will be used."
     ,@mb-url-http-curl-switches))
 
 (defun mb-url-http-sentinel--curl (proc evt)
-  "Curl return the proxy response before the actual remote server response.
+  "Sentinel for Curl.
+
+Curl return the proxy response before the actual remote server response.
 It makes Emacs hard to parse the response message.  Delete the proxy response
 first.
 
@@ -353,6 +389,62 @@ own sentinel instead."
               url-request-extra-headers)
     ,@mb-url-http-httpie-switches))
 
+(defcustom mb-url-http-httpie-supported-content-encoding-list '("gzip" "deflate")
+  "Content encodings which HTTPie supports to decode.
+
+If your HTTPie supports to decode a encoding like br, put the encoding into
+this list.
+
+By default, HTTPie supports to decode gzip and deflate."
+  :type '(repeat (string) :tag "Content Encoding List")
+  :group 'mb-url)
+
+(defun mb-url-http-httpie-delete-content-encoding-from-list (encodings)
+  "Delete \"Content-Encoding\" if all ENCODINGS are supported by HTTPie.
+
+ENCODINGS is a comma-separated string which contains all values of
+\"Content-Encoding\".
+
+Supported encodings are in `mb-url-http-httpie-supported-content-encoding-list'."
+  (cond ((null encodings)
+         t)
+        ((seq-every-p
+          (lambda (encoding)
+            (member encoding mb-url-http-httpie-supported-content-encoding-list))
+          (mapcar 'string-trim (split-string encodings ",")))
+         nil)
+        (t t)))
+
+(defcustom mb-url-http-httpie-content-encoding-fix-function
+  'mb-url-http-httpie-delete-content-encoding-from-list
+  "A function to fix the value of \"Content-Encoding\" for HTTPie."
+  :type '(function :tag "Function")
+  :group 'mb-url)
+
+(defun mb-url-http-sentinel--httpie (proc evt)
+  "Sentinel for HTTPie.
+
+Starting from Emacs 27, `url-handle-content-transfer-encoding' deletes
+\"Content-Encoding\" after decompressing the response body.  HTTPie also
+decodes the response body but it doesn't remove the \"Content-Encoding\"
+header.  This confuses `url-handle-content-transfer-encoding' and
+`url-store-in-cache'.  Delete the header if it exists.
+
+PROC is the process.
+
+EVT describes the type of event."
+  (when (string= evt "finished\n")
+    (with-current-buffer (process-buffer proc)
+      (save-excursion
+        (mb-url-http--delete-carriage-return)
+        (when (>= emacs-major-version 27)
+          (mb-url-http--fix-header "Content-Encoding"
+                                   mb-url-http-httpie-content-encoding-fix-function
+                                   nil nil nil))
+        (mb-url-http--reset-end-of-headers)
+        (goto-char (point-min))
+        (url-http-end-of-document-sentinel proc evt)))))
+
 ;;;###autoload
 (defun mb-url-http-httpie (name url buffer default-sentinel)
   "HTTPie backend for `mb-url-http'.
@@ -367,7 +459,7 @@ DEFAULT-SENTINEL is the default sentinel of mb-url."
   (mb-url-http-make-pipe-process
    url name buffer
    (mb-url-http--httpie-command-list url)
-   default-sentinel))
+   #'mb-url-http-sentinel--httpie))
 
 (provide 'mb-url-http)