+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added ###
 
 - Test for Emacs 28.1
+- HTTPie sentinel
+- Custom variables for deleting HTTPie supported content encodings
 
 ### Changed ###